fix: telemetry message shows extension source
